--- src/shaders/mod.rs
@@ -1,3 +1,0 @@
-include!("interface_fs.rs");
-include!("interface_vs.rs");
-include!("square_vs.rs");
